refactor: simplify WorkScheduler architecture
