+++ logic/src/androidMain/kotlin/com/wire/kalium/logic/sync/WorkSchedulerImpl.kt
@@ -1,30 +1,15 @@
 package com.wire.kalium.logic.sync
 
-import android.app.Notification
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.content.Context
-import android.content.Context.NOTIFICATION_SERVICE
-import android.os.Build
-import androidx.annotation.RequiresApi
 import androidx.work.Constraints
-import androidx.work.CoroutineWorker
-import androidx.work.Data
 import androidx.work.ExistingPeriodicWorkPolicy
 import androidx.work.ExistingWorkPolicy
-import androidx.work.ForegroundInfo
-import androidx.work.ListenableWorker
 import androidx.work.NetworkType
 import androidx.work.OneTimeWorkRequest
 import androidx.work.OutOfQuotaPolicy
 import androidx.work.PeriodicWorkRequest
 import androidx.work.WorkManager
-import androidx.work.WorkerFactory
-import androidx.work.WorkerParameters
-import com.wire.kalium.logic.CoreLogic
-import com.wire.kalium.logic.R
 import com.wire.kalium.logic.data.user.UserId
-import com.wire.kalium.logic.kaliumLogger
 import kotlinx.datetime.Clock
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.Instant
@@ -33,238 +18,95 @@ import kotlinx.datetime.TimeZone
 import kotlinx.datetime.plus
 import kotlinx.datetime.toInstant
 import kotlinx.datetime.toLocalDateTime
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 import java.util.concurrent.TimeUnit
-import kotlin.reflect.KClass
 
-class WrapperWorker(private val innerWorker: DefaultWorker, appContext: Context, params: WorkerParameters) :
-    CoroutineWorker(appContext, params) {
+private val workerClass = WrapperWorker::class.java
+internal actual class GlobalWorkSchedulerImpl(
+    private val appContext: Context
+) : GlobalWorkScheduler {
 
-    override suspend fun doWork(): Result = when (innerWorker.doWork()) {
-        is com.wire.kalium.logic.sync.Result.Success -> {
-            Result.success()
-        }
-        is com.wire.kalium.logic.sync.Result.Failure -> {
-            Result.failure()
-        }
-        is com.wire.kalium.logic.sync.Result.Retry -> {
-            Result.retry()
-        }
-    }
+    override fun schedulePeriodicApiVersionUpdate() {
+        val inputData = WrapperWorkerFactory.workData(UpdateApiVersionsWorker::class)
 
-
-    //TODO(ui-polishing): Add support for customization of foreground info when doing work on Android
-    override suspend fun getForegroundInfo(): ForegroundInfo {
-        val notification = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            Notification.Builder(applicationContext, createNotificationChannel().id)
-        } else {
-            Notification.Builder(applicationContext)
-        }.setContentTitle(NOTIFICATION_TITLE)
-            .setSmallIcon(R.mipmap.ic_launcher) //TODO(ui-polishing): Customize icons too
+        val connectedConstraint = Constraints.Builder()
+            .setRequiredNetworkType(NetworkType.CONNECTED)
             .build()
-        return ForegroundInfo(NOTIFICATION_ID, notification)
-    }
 
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun createNotificationChannel(): NotificationChannel {
-        //TODO(ui-polishing): Internationalis(z)ation. Should come as a
-        //                    side-effect of enabling customization of notifications by consumer apps
-        val name = "Wire Sync"
-        val descriptionText = "Updating conversations and contact information"
-        val importance = NotificationManager.IMPORTANCE_NONE
-        val channel = NotificationChannel(CHANNEL_ID, name, importance)
-        channel.description = descriptionText
-        val notificationManager = applicationContext.getSystemService(NOTIFICATION_SERVICE) as NotificationManager
-        notificationManager.createNotificationChannel(channel)
-        return channel
+        val scheduledHourOfDayToExecute = TIME_OF_EXECUTION // schedule at 4AM
+        val repeatIntervalInHours = REPEAT_INTERVAL // execute every 24 hours
+        val localTimeZone = TimeZone.currentSystemDefault()
+        val timeNow: Instant = Clock.System.now() // current time
+        val timeScheduledToExecute = timeNow.toLocalDateTime(localTimeZone) // time at which the today's execution should take place
+            .let { localDateTimeNow ->
+                LocalDateTime(
+                    localDateTimeNow.year, localDateTimeNow.monthNumber, localDateTimeNow.dayOfMonth,
+                    scheduledHourOfDayToExecute, 0, 0, 0
+                ).toInstant(localTimeZone)
+            }
+        val initialDelayMillis = // delay calculated as a difference between now and next scheduled execution
+            if (timeScheduledToExecute > timeNow) (timeScheduledToExecute - timeNow).inWholeMilliseconds
+            else (timeScheduledToExecute.plus(1, DateTimeUnit.DAY, localTimeZone) - timeNow).inWholeMilliseconds
+
+        val requestPeriodicWork = PeriodicWorkRequest.Builder(workerClass, repeatIntervalInHours.toLong(), TimeUnit.HOURS)
+            .setInitialDelay(initialDelayMillis, TimeUnit.MILLISECONDS)
+            .setConstraints(connectedConstraint)
+            .setInputData(inputData)
+            .build()
+
+        WorkManager.getInstance(appContext).enqueueUniquePeriodicWork(
+            "${UpdateApiVersionsWorker.name}-periodic",
+            ExistingPeriodicWorkPolicy.KEEP,
+            requestPeriodicWork
+        )
     }
 
     private companion object {
-        const val NOTIFICATION_TITLE = "Wire is updating"
-        const val NOTIFICATION_ID = -778899
-        const val CHANNEL_ID = "kaliumWorker"
+        const val TIME_OF_EXECUTION = 4
+        const val REPEAT_INTERVAL: Long = 24
     }
 }
 
-class WrapperWorkerFactory(private val coreLogic: CoreLogic) : WorkerFactory() {
+internal actual class UserSessionWorkSchedulerImpl(
+    private val appContext: Context,
+    override val userId: UserId
+) : UserSessionWorkScheduler {
 
-    override fun createWorker(appContext: Context, workerClassName: String, workerParameters: WorkerParameters): ListenableWorker? {
-        if (WrapperWorker::class.java.canonicalName != workerClassName) {
-            return null // delegate to default factory
-        }
-
-        val userId = workerParameters.getSerializable<UserId>(USER_ID_KEY)
-        val innerWorkerClassName = workerParameters.inputData.getString(WORKER_CLASS_KEY)
-            ?: throw IllegalArgumentException("No worker class name specified")
-
-        kaliumLogger.v("WrapperWorkerFactory, creating worker for class name: $innerWorkerClassName")
-        return when (innerWorkerClassName) {
-            PendingMessagesSenderWorker::class.java.canonicalName -> {
-                require(userId != null) { "No user id specified" }
-                createPendingMessageSenderWorker(workerParameters, userId, appContext)
-            }
-            SlowSyncWorker::class.java.canonicalName -> {
-                require(userId != null) { "No user id specified" }
-                createSlowSyncWorker(workerParameters, userId, appContext)
-            }
-            UpdateApiVersionsWorker::class.java.canonicalName -> {
-                createApiVersionCheckWorker(workerParameters, appContext)
-            }
-            else -> {
-                kaliumLogger.d("No specialized constructor found for class $innerWorkerClassName. Default constructor will be used")
-                createDefaultWorker(innerWorkerClassName, appContext, workerParameters)
-            }
-        }
-    }
-
-    private fun createApiVersionCheckWorker(workerParameters: WorkerParameters, appContext: Context): WrapperWorker {
-        val worker = coreLogic.globalScope {
-            UpdateApiVersionsWorker(updateApiVersions)
-        }
-        return WrapperWorker(worker, appContext, workerParameters)
-    }
-
-    private fun createSlowSyncWorker(workerParameters: WorkerParameters, userId: UserId, appContext: Context): WrapperWorker {
-        val worker = SlowSyncWorker(coreLogic.getSessionScope(userId))
-        return WrapperWorker(worker, appContext, workerParameters)
-    }
-
-    private fun createPendingMessageSenderWorker(workerParameters: WorkerParameters, userId: UserId, appContext: Context): WrapperWorker {
-        val userScope = coreLogic.getSessionScope(userId)
-        val worker = PendingMessagesSenderWorker(
-            userScope.messages.messageRepository,
-            userScope.messages.messageSender,
-            userId
-        )
-        return WrapperWorker(worker, appContext, workerParameters)
-    }
-
-    private fun createDefaultWorker(
-        innerWorkerClassName: String,
-        appContext: Context,
-        workerParameters: WorkerParameters
-    ): WrapperWorker {
-        val constructor = Class.forName(innerWorkerClassName).getDeclaredConstructor()
-        val innerWorker = constructor.newInstance()
-        return WrapperWorker(innerWorker as DefaultWorker, appContext, workerParameters)
-    }
-
-    internal companion object {
-        private const val WORKER_CLASS_KEY = "worker_class"
-        internal const val USER_ID_KEY = "user-id-worker-param"
-        // TODO: delete not used anymore
-        internal const val SERVER_CONFIG_ID_KEY = "server-config-id-worker-param"
-
-        fun workData(work: KClass<out DefaultWorker>, userId: UserId? = null) = Data.Builder()
-            .putString(WORKER_CLASS_KEY, work.java.canonicalName)
-            .apply { userId?.let { putSerializable(USER_ID_KEY, it) } }
-            .build()
-    }
-}
-
-actual sealed class WorkSchedulerImpl(private val appContext: Context) : WorkScheduler {
-    internal val workerClass = WrapperWorker::class.java
-
-    /**
-     * Schedules some work to be done in the background, in a "run and forget" way – free from client app observation.
-     * On mobile clients for example, aims to start a job that will not be suspended by the user minimizing the app.
-     */
-    override fun enqueueImmediateWork(work: KClass<out DefaultWorker>, name: String) {
-        val inputData = WrapperWorkerFactory.workData(work, if (this is UserSession) this.userId else null)
+    override fun enqueueSlowSyncIfNeeded() {
+        val inputData = WrapperWorkerFactory.workData(SlowSyncWorker::class, userId)
 
         val request = OneTimeWorkRequest.Builder(workerClass)
             .setExpedited(OutOfQuotaPolicy.RUN_AS_NON_EXPEDITED_WORK_REQUEST)
             .setInputData(inputData).build()
 
         WorkManager.getInstance(appContext).beginUniqueWork(
-            name,
-            ExistingWorkPolicy.REPLACE,
+            WORK_NAME_SLOW_SYNC,
+            ExistingWorkPolicy.KEEP, // If it's ongoing, let it be
             request
         ).enqueue()
     }
 
-    actual class Global(
-        private val appContext: Context
-    ) : WorkSchedulerImpl(appContext), GlobalWorkScheduler {
+    override fun scheduleSendingOfPendingMessages() {
+        val inputData = WrapperWorkerFactory.workData(PendingMessagesSenderWorker::class, userId)
 
-        override fun schedulePeriodicApiVersionUpdate() {
-            val inputData = WrapperWorkerFactory.workData(UpdateApiVersionsWorker::class)
+        val connectedConstraint = Constraints.Builder()
+            .setRequiredNetworkType(NetworkType.CONNECTED)
+            .build()
 
-            val connectedConstraint = Constraints.Builder()
-                .setRequiredNetworkType(NetworkType.CONNECTED)
-                .build()
+        val request = OneTimeWorkRequest.Builder(workerClass)
+            .setConstraints(connectedConstraint)
+            .setExpedited(OutOfQuotaPolicy.RUN_AS_NON_EXPEDITED_WORK_REQUEST)
+            .setInputData(inputData)
+            .build()
 
-            val scheduledHourOfDayToExecute = TIME_OF_EXECUTION // schedule at 4AM
-            val repeatIntervalInHours = REPEAT_INTERVAL // execute every 24 hours
-            val localTimeZone = TimeZone.currentSystemDefault()
-            val timeNow: Instant = Clock.System.now() // current time
-            val timeScheduledToExecute = timeNow.toLocalDateTime(localTimeZone) // time at which the today's execution should take place
-                .let { localDateTimeNow ->
-                    LocalDateTime(
-                        localDateTimeNow.year, localDateTimeNow.monthNumber, localDateTimeNow.dayOfMonth,
-                        scheduledHourOfDayToExecute, 0, 0, 0
-                    ).toInstant(localTimeZone)
-                }
-            val initialDelayMillis = // delay calculated as a difference between now and next scheduled execution
-                if (timeScheduledToExecute > timeNow) (timeScheduledToExecute - timeNow).inWholeMilliseconds
-                else (timeScheduledToExecute.plus(1, DateTimeUnit.DAY, localTimeZone) - timeNow).inWholeMilliseconds
-
-            val requestPeriodicWork = PeriodicWorkRequest.Builder(workerClass, repeatIntervalInHours.toLong(), TimeUnit.HOURS)
-                .setInitialDelay(initialDelayMillis, TimeUnit.MILLISECONDS)
-                .setConstraints(connectedConstraint)
-                .setInputData(inputData)
-                .build()
-
-            WorkManager.getInstance(appContext).enqueueUniquePeriodicWork(
-                "${UpdateApiVersionsWorker.name}-periodic",
-                ExistingPeriodicWorkPolicy.KEEP,
-                requestPeriodicWork
-            )
-        }
-
-        private companion object {
-            const val TIME_OF_EXECUTION = 4
-            const val REPEAT_INTERVAL: Long = 24
-        }
+        WorkManager.getInstance(appContext).enqueueUniqueWork(
+            WORK_NAME_PREFIX_PER_USER + userId.value,
+            ExistingWorkPolicy.APPEND,
+            request
+        )
     }
 
-    actual class UserSession(
-        private val appContext: Context,
-        override val userId: UserId
-    ) : WorkSchedulerImpl(appContext), UserSessionWorkScheduler {
-
-        override fun scheduleSlowSync() {
-            enqueueImmediateWork(SlowSyncWorker::class, SlowSyncWorker.name)
-        }
-
-        override fun scheduleSendingOfPendingMessages() {
-            val inputData = WrapperWorkerFactory.workData(PendingMessagesSenderWorker::class, userId)
-
-            val connectedConstraint = Constraints.Builder()
-                .setRequiredNetworkType(NetworkType.CONNECTED)
-                .build()
-
-            val request = OneTimeWorkRequest.Builder(workerClass)
-                .setConstraints(connectedConstraint)
-                .setExpedited(OutOfQuotaPolicy.RUN_AS_NON_EXPEDITED_WORK_REQUEST)
-                .setInputData(inputData)
-                .build()
-
-            WorkManager.getInstance(appContext).enqueueUniqueWork(
-                WORK_NAME_PREFIX_PER_USER + userId.value,
-                ExistingWorkPolicy.APPEND,
-                request
-            )
-        }
-
-        private companion object {
-            const val WORK_NAME_PREFIX_PER_USER = "scheduled-message-"
-        }
+    private companion object {
+        const val WORK_NAME_PREFIX_PER_USER = "scheduled-message-"
+        const val WORK_NAME_SLOW_SYNC = "slow-sync"
     }
 }
-
-private inline fun <reified T> Data.Builder.putSerializable(key: String, value: T) = putString(key, Json.encodeToString(value))
-private inline fun <reified T> WorkerParameters.getSerializable(key: String): T? =
-    inputData.getString(key)?.let { Json.decodeFromString(it) }
