+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/SlowSyncScheduler.kt
@@ -1,12 +1,14 @@
 package com.wire.kalium.logic.sync
 
 /**
- * Responsible for [scheduleSlowSync].
+ * Responsible for [enqueueSlowSyncIfNeeded].
  */
 interface SlowSyncScheduler {
 
     /**
-     *  Schedules an execution of [SlowSyncWorker].
+     * Schedules Slow Sync to be done in the background, in a "run and forget" way – free from client app observation.
+     * If already scheduled or running, no need to re-run it.
+     * On mobile clients for example, aims to start a job that will not be suspended by the user minimizing the app.
      */
-    fun scheduleSlowSync()
+    fun enqueueSlowSyncIfNeeded()
 }
